feat: tuning preloading logic
